backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
python-decouple==3.8
python-dateutil==2.9.0
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
psycopg[binary]==3.2.4
dj-database-url==2.1.0
celery==5.4.0
redis==5.0.1
reportlab==4.1.0
urllib3<2
chardet<6
argon2-cffi==23.1.0
